[scripts/owner] Find all target/files references
Extend `suggest_owners.py` to accept an `--all_refs` argument to search
for references to all targets (not just the top-level targets), and to
all files in the project if no target references were found.

This is useful for finding owners for projects that either don't have a
top-level gn target, or are not built with gn at all and are referenced
directly by file instead.

Change-Id: I56c4932d4c28c4da096623d9d5736d7f9b6f6842
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684562
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Chris Palmer <palmer@google.com>
Commit-Queue: Marina Ciocea <marinaciocea@google.com>