[ffx] Move ffx emu commands to a common directory
As we start to expand the set of operations available on an emulator via
ffx emu, the source code tree will become more and more difficult to
navigate.

This CL moves the existing subcommands into a common "commands"
directory, and adjusts the BUILD.gn files accordingly. There is no
functional change in this CL.

Change-Id: I18d76bf565a31ce0856b4b81b1c3c796b724f8cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684302
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>