[ffx][config] Don't call config.save() twice.
As-is, the 'else' path of this function will always be called, even
if there's a build_dir set. To make it not call it, map_or_else
would need to be used. But I think it's clearer to just only
initialize the build_dir and then use it.

Change-Id: I96ea2d7e401d32a9e56fa1378d01980669ff7e64
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685324
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>