[component][fuzzing] Fix mutation_depth
This CL addresses an issue in which the number of mutations was not
being incremented by realm fuzzing engine.

In the process of fixing that issue, this CL revealed a second bug:
seed corpus elements were not being included in the live corpus and
therefore never picked for mutation. This CL also addresses that issue,
and includes a new method and unit tests for adding one corpus to
another.

Finally, this CL tweaks the libFuzzer runner tests in order to make
coverage more predictable in testing.

Change-Id: I6b78258a3e78c718f41de3933df9e0032895dce9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681144
Reviewed-by: Dmitry Yatsushkevich <dmitryya@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>