[assembly] Configure example based on build type.
After discussion it seems like it might be overkill for the example
to have its own platform config, so we'll infer it from the build type
as a demontration for the more common case where people will be doing
that.

Deletes two tests which checked to make sure that the combination of
platform config and experimental ffx flag worked together correctly.
Now the ffx flag is the only thing which determines whether the example
is configured.

Change-Id: Iae47455c47c045de2034b8a735195d139e90c003
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685576
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>