[fxfs] Change the storage benchmarks to use v2 components
This changes the storage benchmarks so that it runs as a v2 component
which allows us to run Fxfs as a component.  To make things simpler,
this change now runs ODU as a library rather than spawning ODU as
another component.  For now, ODU is the only benchmark that we support.

Bug: 99591
Change-Id: I001aac02687eafa46828b784d249a7eefbb7c156
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684846
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Chris Suter <csuter@google.com>