[fidl] Remove debug assertion that boolean value is 0 or 1
There currently is a debug assertion in the bool WireCodingTraits that
ensures that bool values are 0 or 1 in the encode direction. However,
the reinterpret_cast in this check uses UB.

This check should not be needed unless there is memory corruption
because the bool value is already guaranteed to be 0 or 1. Given this,
this CL removes the unnecessary check.

Fixed: 100689
Change-Id: Ieafc6ae19bc4723e4c6167de1873ad5ea4ccd82e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683679
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>