[fidlc] Add tests for interleaving library availabilities
This CL extends availability_interleaving_tests.cc to test
availabilities on libraries themselves. It also adds a test for the
@available(platform="...") syntax, since the main change incidentally
reduced coverage of that feature.

In particular, the tests reinforce that the following is allowed:

    // foo.fidl
    @available(added=1)
    library foo;
    using bar;
    @available(added=2)
    alias S = bar.S;

    // bar.fidl
    @available(added=2)
    library bar;
    type S = struct {};

It is not an error for 'foo' to import 'bar', even though 'bar' does not
exist at version 1, because 'foo' does not actually use any element of
'bar' until version 2. This is a consequence of treating absent
libraries as empty (I42f528623bccc38385342b7b889d495e242dbb61).

Test: fx test fidl-compiler
Change-Id: Ida907c4400a97f00007fa04598dddb7415b7198b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686110
Fuchsia-Auto-Submit: Mitchell Kember <mkember@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>