[component][fuzzing] Standardize "main" files.
This CL goes through and cleans up all the C++ files in the fuzzing
framework that contain a "main" function, ensuring that they are named
similarly ("-main.cc") and have the same basic layout. This should make
it easier for readers to locate and follow the major entry points.

Change-Id: I197ff5ec8c847051ff514c588fe7eeef4aed2790
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680315
Reviewed-by: Cameron Finucane <eep@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>