[starnix] Allows access to the filesystem withtout a task
Until now, going through the filesystem required having access to a
Task.

It was needed for:
- Resolving symbolic links to absolute path
- Resolving /proc special symbolic links

This prevented accessing the filesystem while working at the runner
level.

This CL move the state state inside the LookupContext and allows it to
be empty, so that special context can be built and used when traversing
the file system outside of the context of a task.

Change-Id: I91475134cc9e9698c7e2780ac55075656d4c0cc4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686223
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Barth <abarth@google.com>