[ffx] Add console sockets to the flags template
In the attached bug, we have a request to enable the user/ffx emu tool
to communicate with the emulator backend through the use of Linux
Sockets. The QEMU backend supports attaching to sockets for several
devices; most importantly for this work are the QEMU Monitor console,
the QMP (a json-based command processor), and the serial port (which
exposes the Fuchsia console).

This CL adds the appropriate flags to the template file for attaching
sockets to these consoles. Tools to utilize the sockets will be
developed in later CLs.

Bug: 94232
Change-Id: Iae54697c3f9b6872c47d42674e32529d955c3413
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682465
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>