[kernel][benchmarks] Don't trigger lockup detector during `k bench`
When running in-kernel benchmarks (`k bench`) that disable interrupts
for extended periods, mark the CPU as inactive so that the lockup
detector knows that it's OK for it to not update it's heartbeats.  The
lockup detector only monitors active CPUs.

Test: Tested by running `k bench` on QEMU and observing no OOPS.
Fixed: 101553
Change-Id: I14956f85ee7f15c700d9b42afa8e178ae738cadd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686322
Fuchsia-Auto-Submit: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Gurjant Kalsi <gkalsi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>