[build][assembly] Only allow the fuchsia toolchian in //build/images
Stumbled across this by accident.  This change adds an assert that the
//build/images/BUILD.gn file is only processed in the Fuchsia toolchain,
not the host (or any other variant), as well as correcting the one place
that was causing it to be processed in the host toolchain.

Change-Id: I6fc1b6562e02dea736b642525829e579a417d5e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686106
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>