[debugger] Remove support for Component Root
Component root is a concept in CFv1 design and it's no longer relevant
in CFv2 any more. It's also broken after the debug_agent is migrated to
CFv2. This patch removes the support completely. The protocol version
is intentionally not bumped because no one is using this feature and no
incompatibility should be introduced.

This patch also includes some miscellaneous changes to improve logging
and docs.

Change-Id: Ic2f5bbfa67a2900f9994a29798d2a937019d22dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684354
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Dangyi Liu <dangyi@google.com>