[netstack3] Deliver multicast UDP to listeners
Deliver incoming UDP packets with multicast destination addresses to all
matching sockets. This is currently limited to listener sockets, since
it is not possible to create a connected socket that is bound to a local
multicast address. Support for creating connected sockets bound to a
multicast address will be added in a future commit.

Bug: 39479
Change-Id: I3408821650fcaf894fa46aa9bd6f404cc9f892b0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682943
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>