[modular][cleanup] Delete basemgr_launcher
basemgr_launcher is used to launch basemgr as a v1 component with a
custom Modular configuration.

This is no longer supported. basemgr should be launched as a v2
component, and any changes to Modular config should be made to
the configuration bundled in the basemgr package.

Change-Id: I3b4377d907fe08dc331caa9850a6462862c49171
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686573
Reviewed-by: Gabe Schine <thatguy@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>