[rfc] Amend RFC-0103 to provide rationale.
RFC-103 defined the situations in which SWD changes will require an RFC.
During recent discussion of an in-work SWD RFC it became clear that the
rationale for requiring an RFC was not clear in some of these,
specifically the meaning and cost of a stepping stone release was not
clear to some readers.

This ammendment adds a short justification for each of the situations in
which a SWD RFC is required.

Change-Id: I2de011a011e06fd9f75632e735cad99bc12a60d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685372
Reviewed-by: John Wittrock <wittrock@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Olivia Nordquist <nolivia@google.com>