[netcfg] Allow forwarding based on device class
This change adds a configuration section to allow forwarding on a device
based on its device class.

A new integration test checks both the positive and negative cases of
interfaces having a configuration installed based on the class. The
previous API implied that extensions to `KnownServiceProvider::Manager`
tests would require additional arguments, but this didn't appear to
scale well in the actual implementation. That's modified to now check
the used config, with the assumption that tests target specific
configurations. This can be improved if netcfg later is able to receive
a configuration as a FIDL message.

Fixed: 94461
Change-Id: Ib34cda8025101ccf6f3e06e6b88def5075892126
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682947
Commit-Queue: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>