[a11y] Use the new 'entered list detail' TTS.
(instead of concatenating together one TTS for 'entered list' and one
for 'with N items', which was causing an uncanny pause.)

This CL also fixes a typo in the EnterAndExitListsAndTables test. It
didn't affect the correctness of the test, but might have been
misleading.

Change-Id: I1bd38269b3042407f22d435078d6e27b687ccfdd
Fixed: 99346
Test: fx test a11y-lib-tests
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/676330
Reviewed-by: Alexander Brusher <abrusher@google.com>
Commit-Queue: Michael LoPiccolo <milopiccolo@google.com>