[scripts/owner] Generate OWNERS files
Extend `suggest_owners.py` to accept an `--generate_missing` argument to
generate OWNERS files for the projects that are missing owners.

Each generated OWNERS file includes the OWNERS files of all the projects
that reference the project.

For now, the OWNERS files are created in the working directory, to be
uploaded to CQ in a follow-up change.

Change-Id: I8b75d257d7f6761ed49c2f9613ca3bcaf186e6a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686422
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Chris Palmer <palmer@google.com>
Commit-Queue: Marina Ciocea <marinaciocea@google.com>