[gigaboot] UTF conversion doc fixes and tests
Fix the UTF conversion function docs to indicate that the conversion
does *not* append null termination, and add some 16-to-8 conversion
unittests.

No functionality change, this part of the code doesn't seem to be
causing the linked commandline bug, but as long as we're investigating
this code anyway, adding unittests seems like a good idea.

Bug: 100791
Change-Id: I76f68b246e89f007c22ada16616e0761e83a050a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686500
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: David Pursell <dpursell@google.com>