[ffx] Remove autoconnect, unify socket logic.
This change unifies the logic around where to find the ascendd
socket by placing the responsibility completely on the daemon
spawner and client (ffx) to determine it, and has it using
either the environment variable if present or the config value
if not.

As part of this, the autoconnect logic is completely removed.
It appeared to be disabled everywhere anyways, so just get rid
of it and if we want to add it back in elsewhere we can still
do that, but we should do it better (and with no or less reliance
on a lazy_static to initialize a connection).

This may be related to the sporadic errors in 94564, but that's
unclear so far.

Bug: 94564
Fixes: 64499
Change-Id: I1114dbe269a019df4ef909cb33ca0a939a3409b3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685846
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>