[assembly] Create 'minimal' product
Factor items out of core.gni to create a "minimum viable" base for a
Fuchsia product, then layer the 'core' product definition on top of
that.

This is not a completed effort, there will be follow-on commits to
move additional packages between these configs as we iterate.

Bug: 100671

Change-Id: Ic0ff842d00642835274d2e647575fcdcae5e5766
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683826
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Chris Masone <cmasone@google.com>