[recovery] Do not "re"-load default FontFace.
A request to stop "re"-loading default font was made in fxr/666557.

This CL causes RecoveryAppAssistant's create_view_assistant() to call
load_default_font_face() once.

Copies of the loaded FontFace are passed to ConsoleViewAssistant and
RecoveryViewAssistant's constructors.

Now, instead of loading the Font when rendering a Console or Recovery
view, the "local copies" are used.

For Reviewer: looking for Rust guidance on "It'd be smart to use a
Singleton from Google-friendly crate $XYZ" or "Too much cloning of
loaded font is going on, refactor this to use a ref to a single
object."

If this approach is reasonable, I'll update KeyboardViewAssistant
in a follow-up CL.

Bug: b/233983593
Tested: fx test system_recovery_bin_test
Change-Id: I0b2b1c309eddc69ca25187709f37db460e921164
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685424
Reviewed-by: Kevin Twidle <kpt@google.com>
Commit-Queue: Mike McCreavy <mccreavy@google.com>