[brcmfmac][wlan] Return frames on TX fail
In the event that brcmfmac failed to queue up TX frames the frames were
not properly returned to netdevice. This requires the priority queue to
change slightly as it's not safe to handle a moved-from frame if
queueing fails. It also doesn't cover the case where a lower priority
frame queue is evicted, the evicted frame also needs to be returned.

Solve this by providing a pointer to a Frame in the push call where the
evicted or failed frame is placed if needed. This does come at the cost
of an allocation when this occurs. But because we expected users to
provide the TX queue depth to netdevice it is extremely rare that the
queue would overflow so it's better to pay this price only on failure
while reducing the cost as much as possible for the normal use case.
Handle this properly in brcmfmac as well as cases where the entire data
path is down or the SDIO bus is down that would prevent transmitting
frames.

Bug: 101156
Test: Manually ensure no buffer leakage by decreasing TX queue size and
observing queue full events. Included unit test.

Change-Id: I02d653a09f61931943d8f7c6d5fa9833b080c2c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683822
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Commit-Queue: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>