[ffx] Common functions for deserializing engines
Previously, each emu subcommand lib.rs file would include a function
named something like "get_instance_path", which would look up an
emulator working directory, do some validation, and deserialize an
engine instance from that directory, then the tool would invoke the
actual command on that instance.

This CL factors out the engine instantiation and validation code from
each subcommand and makes it a common library for all the subcommands.
This simplifies the code in all of the subcommand libs.

Although this change will help us get to a point where we can better
test, once everything is behind traits that can be mocked, the code in
this CL does not improve our testing coverage because we don't have a
way to deserialize an engine type other than the two production types
(femu and qemu). This is an ongoing issue tracked in fxbug.dev/96560.

Change-Id: I42b39f90005e9cf72546a0fe24f3a7140cf1eb1f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685331
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>