[ffx] Clean up extra build imports
It was pointed out in fxrev.dev/685331 that the components.gni file is
not being used by any of the libraries in the ffx emu plugin, but is
imported all the same. This CL removes the extra imports for code
cleanliness. There is no functional change.

Change-Id: Icba2cadef1edc3f5c4b34b09a599877ab9c928ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687228
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>