[kernel][dev][power] add a quick reset driver for the AS370 SoC
Largely a resurrection of an existing driver deleted in change
21d77ff05881f8af8a5c6e0f87608abd587e91cd

Attempts to use the watchdog driver to reset the system. This is needed
currently due to a lack of a functional PSCI_SYSTEM_RESET hook in the
current firmware. Hopefully will be made obsolete once that is properly
implemented.

Big: 94099

Change-Id: I9b7bddee3237b7ab69b5c780cc5f8d6ba783076a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686763
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>