[staticlints] Check for doc links pointing to Gitiles
Links to docs from source code should link to the doc in fuchsia.dev,
rather than linking to the .md file in Gitiles, primarily because docs
are much nicer to read in fuchsia.dev than in Gitiles.

Any link of the form
"https://fuchsia.googlesource.com/fuchsia/+/<branch/<path>.md" will be
now flagged with a suggestion to replace it with
"https://fuchsia.dev/fuchsia-src/<path>".

This will just be a Tricium warning and will not block people from
submitting links to docs in Gitiles, which is reasonable behavior since
it's not the end of the world if someone does submit such a link, and
there may be legitimate reasons for doing so.

Bug: b/144128643
Change-Id: Ifb118cd1981b7ee019eb130c756298fc021946a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686445
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ina Huh <ihuh@google.com>