[fidl] Convert driver-unit-test to unified FIDL bindings
Currently driver-unit-test uses the C bindings, which are deprecated.
This CL converts driver-unit-test to use the unified C++ bindings,
which continue to be maintained.

The main motivation for this CL was to get rid of usage of
fidl::Builder, which I hope to remove.

Change-Id: Ie051fa59a958149e506ae18f8a1fcd78993fc69c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687422
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>