[fuchsia-fs] rename io_util to fuchsia-fs
fuchsia-fs is a combination of the fuchsia-component/fuchsia-async
naming convention for core fuchsia rust libraries, as well as the
std::fs module for filesystem operations in the standard library. It's
the new location for standard fuchsia filesystem operations, such as
opening files and directories, using fuchsia primitives.

This also includes a redirection for dependencies on io_util, so no
conversions to the new library are made in this cl, to keep it cleaner.

Bug: 36063
Bug: 61251
Change-Id: I8b221a0c70615b2a3a495abc4393c2e8c7f2fca3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687684
Reviewed-by: Shai Barack <shayba@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Commit-Queue: Stephen Demos <sdemos@google.com>