[build][tracer] Also detect and ignore temp dirs
Workaround issue where fsatrace fails to report mkdirs.
This messes up the logic used to infer temporary files/dirs.
Since we don't care about dirs (as they have no content),
we opt to ignore them.
If a path is currently an existing directory, it is ignored as before.
If a path *was* a directory that no longer exists, ignore it as well.
File paths are still checked as before.

Bug: 101488

Change-Id: I1d810e256b5113bee71172f55ec4ba714f0d652f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686462
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: David Fang <fangism@google.com>