[fastboot] Add is-userspace variable
Add a fastboot variable `is-userspace` that always return `yes` for
userspace fastboot. The plan is to use it in the flash script to
determine whether it is running from userspace fastboot and need to
perform any specific actions such as rebooting after flashing the fvm
image. i.e.:

  fastboot flash fvm <fvm block>
  if (fastboot getvar is-userspace 2>&1 | grep "yes"); then
    fastboot reboot bootloader
  fi

Bug: b/217597389
Change-Id: I745c564558f08140bd900c84f57e7abf4bf8358c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687531
Reviewed-by: David Pursell <dpursell@google.com>
Commit-Queue: Yecheng Zhao <zyecheng@google.com>