[rbe] new rust_rbe structured global variables
Re-organize variables into: rbe_rust.{parameter} for clarity.

The default values mirror the original ones.
Uses of the variable have been updated to (old || new), where the
default is false.  The old variables will be phased out in a soft
transition.  This commit is a non-breaking change.

`rbe_exec_strategy` was replaced in this CL because it is not
referenced outside of the main tree.

Future: rbe_cxx.* for C++

Change-Id: I9338702bd85a537b9626c4deb4aac582b494912b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/659362
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: David Fang <fangism@google.com>