[component_manager] Make CollectionServiceProvider more generic
* Rename CollectionServiceProvider to CollectionCapabilityProvider

This reflects that it provides instances of any capability described
by an ExposeDecl and CapabilityName, not just service capabilities.

It also avoids confusion with service instances, which are not
individual capabilities.

* Accept a CapabilityName rather than a target declaration

The declaration is only used to read the capability source name,
which is static, so the caller can pass it in directly and avoid
having to copy/construct the decl.

Bug: 4776
Change-Id: I4f908a642e0e78bb1d3a2f4f50102acda1aacbeb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687104
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Reviewed-by: Ryan Diaz <ryandiaz@google.com>