[component_manager] Return get_default_provider errors to caller
Ensures that get_default_provider returns an error when it has
a default provider, but can't construct it, rather than logging
and returning None as if it didn't have one.

This makes the routing code fail fast instead of then trying to
find a provider via events, which is unlikely to return anything.

Change-Id: I39607d888a7f39bc1b4f87d67b0044d66cb1c99c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687107
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Ryan Diaz <ryandiaz@google.com>