[ffx][config] Remove magic for tests.
It's much much less confusing to just have tests be clear about how
they're initializing the environment.

This means that if you write a test that uses config and it's run
independently from ffx it will error until you run the test_init()
function, but I think that's better. We could add a wrapper macro
to do this automatically if we really want (and that might be good
because it could also tear it down), but this is a good first step
to that.

Change-Id: I5e7c49f529c3e281bb7b18e4cc7a21cd6b00e748
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/685323
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>