[sys] ServeFromStartupInfo should fail if called twice
Change OutgoingDirectory::Serve and
OutgoingDirectory::ServeFromStartupInfo to fail with ZX_ERR_BAD_HANDLE
when the directory request is not valid.

The other option is to follow go/fuchsia-content and crash with an error
message string. However, the existing documentation of OutgoingDirectory
already documented a behavior for this error case, which is to return an
error code. Doing so seemed to mesh better with existing contracts.

Fixed: 92360
Change-Id: I2564f77ee74f8693dafb75e65e86d8029573ac3d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688142
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>