[netstack3] Don't require InstantContext for IpStateContext
A later change will only implement `InstantContext` on the non-sync
context but `IpStateContext` will be implemented on the sync context.
`IpStateContext` implements `InstantContext` so it has access to
a type providing an implementation of `Instant` but we can get that
with a type paramter on the `IpStateContext` trait instead.

Bug: 48578, 86168

Change-Id: Icf624d3f08cf2cd29c49b98e0064192ef3bcb93c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687663
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Jake Fried <friedj@google.com>
Commit-Queue: Ghanan Gowripalan <ghanan@google.com>