[pi][test] Add a simple profile inheritance stress test.
Add a simple command line program which can be used to stress profile
inheritance in the kernel. The basic idea is to attempt to rapidly
generate many different profile inheritance situations in order to
exercise as many strange edge cases involving profile inheritance in
the kernel as possible.

Note that this is not a stress test.  It does not construct a set of
preconditions, perform a single operation, then rigorously verify a
set of postconditions.  Instead, it relies on ASSERTS and other
runtime debug checks in the kernel to detect violations of invariants
while focusing on producing as many different asynchronous interactions
as possible.  It is meant to be run for long periods of time (days) in
order to flush out rare edge cases, and it has been an important part
of developing the transition from priority inheritance to profile
inheritance.

Change-Id: Ide5f782b34d4b5ab0e486456fe63392a4c9fc871
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/524420
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>